requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
--index-url http://pip.duxiaoman-int.com/root/pypi/+simple/
--trusted-host pip.duxiaoman-int.com

accelerate==1.3.0
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
aiosignal==1.3.2
    # via aiohttp
async-timeout==5.0.1
    # via aiohttp
attrs==25.1.0
    # via aiohttp
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
datasets==3.3.2
    # via -r requirements.in
diffusers==0.33.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.17.0
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.12.0
    # via
    #   datasets
    #   huggingface-hub
    #   lightning
    #   pytorch-lightning
    #   torch
hf-xet==1.1.3
    # via huggingface-hub
huggingface-hub==0.32.4
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   peft
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via diffusers
jinja2==3.1.5
    # via torch
lightning==2.5.2
    # via -r requirements.in
lightning-utilities==0.14.3
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.4.2
    # via torch
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
opencv-python==4.11.0.86
    # via -r requirements.in
opencv-python-headless==4.11.0.86
    # via -r requirements.in
packaging==24.2
pandas==2.2.3
    # via datasets
peft==0.14.0
    # via -r requirements.in
pillow==11.1.0
    # via
    #   diffusers
    #   torchvision
prodigyopt==1.1.2
    # via -r requirements.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via
    #   accelerate
    #   peft
pyarrow==19.0.1
    # via datasets
python-dateutil==2.9.0.post0
    # via pandas
pytorch-lightning==2.5.2
    # via lightning
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   lightning
    #   peft
    #   pytorch-lightning
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   transformers
safetensors==0.5.2
    # via
    #   accelerate
    #   diffusers
    #   peft
    #   transformers
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.6.0
    # via
    #   -r requirements.in
    #   accelerate
    #   lightning
    #   peft
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==1.7.3
torchvision==0.21.0
    # via -r requirements.in
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
wandb
matplotlib
# The following packages are considered to be unsafe in a requirements file:
# setuptools
